+++ lib/td_dd/data_structures/data_structure.ex
@@ -13,6 +13,10 @@ defmodule TdDd.DataStructures.DataStructure do
 
   @behaviour Searchable
 
+  @status %{
+    deleted: "deleted"
+  }
+
   schema "data_structures" do
     belongs_to(:system, System, on_replace: :update)
     has_many(:versions, DataStructureVersion, on_delete: :delete_all)
@@ -30,6 +34,7 @@ defmodule TdDd.DataStructures.DataStructure do
     field(:name, :string)
     field(:ou, :string)
     field(:type, :string)
+    field(:deleted_at, :utc_datetime)
 
     timestamps(type: :utc_datetime)
   end
@@ -41,7 +46,8 @@ defmodule TdDd.DataStructures.DataStructure do
       :confidential,
       :df_content,
       :last_change_at,
-      :last_change_by
+      :last_change_by,
+      :deleted_at
     ])
   end
 
@@ -120,6 +126,12 @@ defmodule TdDd.DataStructures.DataStructure do
       |> Map.get(:system)
       |> (& &1.__struct__.search_fields(&1)).()
 
+    status =
+      case structure.deleted_at do
+        nil -> ""
+        _del_timestamp -> @status.deleted
+      end
+
     %{
       id: structure.id,
       description: structure.description,
@@ -137,11 +149,29 @@ defmodule TdDd.DataStructures.DataStructure do
       inserted_at: structure.inserted_at,
       confidential: structure.confidential,
       df_content: structure.df_content,
-      data_fields: Enum.map(structure.data_fields, &DataField.search_fields/1),
+      data_fields: non_deleted_data_fields(structure),
       path: structure.path,
+      status: status,
       class: structure.class
     }
   end
+
+  defp non_deleted_data_fields(%{data_fields: data_fields}) do
+    data_fields
+    |> Enum.map(&DataStructures.with_field_structure/1)
+    |> Enum.filter(fn data_field -> not is_nil(Map.get(data_field, :field_structure)) end)
+    |> Enum.filter(fn data_field ->
+      deleted_at =
+        data_field
+        |> Map.get(:field_structure, %{})
+        |> Map.get(:deleted_at)
+
+      is_nil(deleted_at)
+    end)
+    |> Enum.map(&DataField.search_fields/1)
+  end
+
+  defp non_deleted_data_fields(_), do: []
 
   defp fill_items(structure) do
     keys_to_fill = [:name, :group, :ou]